+++ src/Models/HoaDonNhapSach.java
@@ -4,6 +4,7 @@
  */
 package Models;
 
+import java.math.BigInteger;
 import java.sql.Date;
 
 /**
@@ -12,7 +13,7 @@ import java.sql.Date;
  */
 public class HoaDonNhapSach {
 
-    private long id;
+    private BigInteger id;
     private String nguoiXuLy;
     private float tongTien;
     private Date ngayTao;
@@ -21,7 +22,7 @@ public class HoaDonNhapSach {
     public HoaDonNhapSach() {
     }
 
-    public HoaDonNhapSach(long id, String nguoiXuLy, float tongTien, Date ngayTao, String qr_code) {
+    public HoaDonNhapSach(BigInteger id, String nguoiXuLy, float tongTien, Date ngayTao, String qr_code) {
         this.id = id;
         this.nguoiXuLy = nguoiXuLy;
         this.tongTien = tongTien;
@@ -29,11 +30,11 @@ public class HoaDonNhapSach {
         this.qr_code = qr_code;
     }
 
-    public long getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
